pyproject.toml
[project]
name = "open-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keyboard>=0.13.5",
    "openai-whisper>=20240930",
    "pillow>=11.2.1",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
]
